pg-finder/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
sentence-transformers
pydantic
langgraph
langchain
langchain-groq
langchain-core
qdrant-client
fastembed
